make category and product

+++ src/main/java/com/example/springboot/controller/CategoryController.java
@@ -1,12 +1,15 @@
 package com.example.springboot.controller;
 
+import com.example.springboot.common.ApiResponse;
 import com.example.springboot.model.Category;
 import com.example.springboot.service.CategoryService;
-import com.example.springboot.service.ProductService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
+import java.util.Optional;
 
 @CrossOrigin(origins = "http://localhost:3000/")
 @RestController
@@ -15,14 +18,16 @@ public class CategoryController {
     @Autowired
     private CategoryService categoryService;
 
-    @PostMapping(path="/add")
-    public @ResponseBody Category addNewProduct (@RequestBody Category category)
+    @PostMapping(path="/create")
+    public @ResponseBody ResponseEntity<ApiResponse> addNewCategory (@RequestBody Category category)
     {
-        return categoryService.saveCategory(category);
+        categoryService.saveCategory(category);
+        return new ResponseEntity<>(new ApiResponse(true, "Thêm Danh Mục Mới Thành Công:" + " " + category.getName()),
+                HttpStatus.CREATED);
     }
 
     @GetMapping(path="/all")
-    public @ResponseBody List<Category> getAllProducts() {
+    public List<Category> getAllCategories() {
         return categoryService.getCategories();
     }
 
@@ -33,16 +38,31 @@ public class CategoryController {
         return categoryService.findBySlug(slug);
     }
 
-    @PutMapping("/edit")
-    public Category editProduct(@RequestBody Category category)
+    @PutMapping(path="/update/{slug}")
+    public @ResponseBody ResponseEntity<ApiResponse> editCategory(@PathVariable String slug, @RequestBody Category category)
     {
-        return categoryService.updateCategory(category);
+        if(categoryService.findBySlug(slug) == null)
+        {
+            return new ResponseEntity<>(new ApiResponse(false, "Không Tìm Thấy Danh Mục Nào Có Slug Là: " + slug),
+                    HttpStatus.NOT_FOUND);
+        }
+        categoryService.updateCategory(slug, category);
+        return new ResponseEntity<>(new ApiResponse(true, "Sửa Thành Công Danh Mục:" + " " + category.getName()),
+                HttpStatus.OK);
     }
 
-    @DeleteMapping("/delete/{id}")
-    public String deleteProduct(@PathVariable int id)
+    @DeleteMapping(path="/delete/{id}")
+    public @ResponseBody ResponseEntity<ApiResponse> deleteCategory(@PathVariable int id)
     {
-        return categoryService.deleteCategory(id);
+        Optional<Category> existCategory = Optional.ofNullable(categoryService.findById(id).orElse(null));
+        if(existCategory.isEmpty())
+        {
+            return new ResponseEntity<>(new ApiResponse(false, "Danh Mục Bạn Xóa Không Tồn Tại"),
+                    HttpStatus.NOT_FOUND);
+        }
+        categoryService.deleteCategory(id);
+        return new ResponseEntity<>(new ApiResponse(true, "Xóa Thành Công Danh Mục:" + " " + existCategory.get().getName()),
+                HttpStatus.OK);
     }
 
 }
